Revert "Deprecated React, began PHANTOM buiild"
