+++ legacy/synth.tsx
