+++ package.json
@@ -18,6 +18,7 @@
     "@babel/core": "^7.8.3",
     "@babel/plugin-proposal-class-properties": "^7.8.3",
     "@babel/preset-env": "^7.8.3",
+    "@babel/preset-react": "^7.8.3",
     "@typescript-eslint/eslint-plugin": "^2.19.0",
     "@typescript-eslint/parser": "^2.19.0",
     "babel-loader": "^8.0.0-beta",
@@ -28,6 +29,7 @@
     "eslint-loader": "^3.0.3",
     "file-loader": "^5.0.2",
     "html-webpack-plugin": "^3.2.0",
+    "react-helmet": "^5.2.1",
     "style-loader": "^1.1.2",
     "uglifyjs-webpack-plugin": "^1.0.0-rc.0",
     "webpack": "^4.41.5",
@@ -40,6 +42,9 @@
   "dependencies": {
     "@emotion/core": "^10.0.27",
     "@emotion/styled": "^10.0.27",
+    "@types/react": "^16.9.19",
+    "@types/react-helmet": "^5.0.15",
+    "@types/react-redux": "^7.1.7",
     "@types/three": "^0.103.2",
     "audiosynth": "^0.0.5",
     "browser-monads": "^1.0.0",
@@ -49,7 +54,12 @@
     "eslint-plugin-prettier": "^3.1.2",
     "gh-pages": "^2.2.0",
     "prettier": "^1.19.1",
-    "redux": "^4.0.5",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-reveal": "^1.2.2",
+    "react-three-fiber": "^4.0.8",
+    "spotify-sdk": "^1.0.0-beta2",
+    "svg-react-loader": "^0.4.6",
     "three": "^0.116.1",
     "ts-loader": "^6.2.1",
     "typescript": "^3.7.5",
