+++ src/components/Logotype.tsx
