+++ src/index.html
@@ -1,9 +1,4 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <!--  
+<!--  
       .-----. 
      /  .---.\    
     /  / _  _ \\  JUST DO SH*T.
@@ -17,15 +12,23 @@
     TOKYO—BASED DVLPMNT 'N' DSGN
     vic@sidiousvic.dev 
     
-    -->
+-->
+
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta property="og:title" content="SIDIOUSVIC.DEV" />
     <meta property="og:image" content="" />
     <meta
       property="og:description"
       content="Axewielding, inditing, programming primate of the species mēxihcan 🍕"
     />
-    <title>🔥 🌻 🍕 👽 😈 🐊</title>
+    <title>🔥 🌻 🍕 👽 🐍 😈 🐊 ⛩ 🏀</title>
     <link rel="icon" type="image/png" href="./static/favicon.png" id="icon" />
   </head>
-  <body></body>
+  <body>
+    <div id="root"></div>
+  </body>
 </html>
