+++ src/modules.d.ts
@@ -1,50 +1,50 @@
 declare module "*.obj" {
-  import SidiousSkullModel from "./assets/models/SidiousSkull.obj";
+  import SidiousSkullModel from "./models/SidiousSkull.obj";
   export default SidiousSkullModel;
 }
 
 declare module "*.ico" {
-  import favicon from "./assets/images/favicon.ico";
+  import favicon from "./images/favicon.ico";
   export default favicon;
 }
 
 declare module "*.svg" {
-  import BlmSvg from "./assets/images/blmSvg.svg";
+  import BlmSvg from "./images/blmSvg.svg";
   export default BlmSvg;
 }
 
 declare module "*.svg" {
-  import XicuahuaBorder from "./assets/images/xicuahua.svg";
+  import XicuahuaBorder from "./images/xicuahua.svg";
   export default XicuahuaBorder;
 }
 
 declare module "*.jpg" {
-  import BlackwatersBackground from "./assets/images/blackwaters.jpg";
+  import BlackwatersBackground from "./images/blackwaters.jpg";
   export default BlackwatersBackground;
 }
 
 declare module "*.png" {
-  import qrcodeScars from "./assets/images/qrcodeScars.png";
+  import qrcodeScars from "./images/qrcodeScars.png";
   export default qrcodeScars;
 }
 
 declare module "*.otf" {
-  import Nikes from ".assets/fonts/FuturaCondensedExtraBoldOblique.otf";
+  import Nikes from "./fonts/FuturaCondensedExtraBoldOblique.otf";
   export default Nikes;
 }
 
 declare module "*.otf" {
-  import NeueMachina from ".assets/fonts/NeueMachina-Regular.otf";
+  import NeueMachina from "./fonts/NeueMachina-Regular.otf";
   export default NeueMachina;
 }
 
 declare module "*.otf" {
-  import NeueMachinaUltrabold from ".assets/fonts/NeueMachina-Ultrabold.otf";
+  import NeueMachinaUltrabold from "./fonts/NeueMachina-Ultrabold.otf";
   export default NeueMachinaUltrabold;
 }
 
 declare module "*.ttf" {
-  import Flag from "./assets/fonts/Flag.ttf";
+  import Flag from "./fonts/Flag.ttf";
   export default Flag;
 }
 
@@ -52,4 +52,3 @@ declare module "browser-monads";
 declare module "react-reveal/Fade";
 declare module "webfontloader";
 declare module "react-dom";
-declare module "diff";
