--- src/styles/fonts.css
@@ -1,17 +1,0 @@
-@font-face {
-  font-family: "Nikes";
-  src: url("../assets/fonts/FuturaCondensedExtraBoldOblique.otf")
-    format("opentype");
-}
-@font-face {
-  font-family: "NeueMachina";
-  src: url("../assets/fonts/NeueMachina-Regular.otf") format("opentype");
-}
-@font-face {
-  font-family: "NeueMachinaUltrabold";
-  src: url("../assets/fonts/NeueMachina-Ultrabold.otf") format("opentype");
-}
-@font-face {
-  font-family: "Flag";
-  src: url("../assets/fonts/Flag.ttf") format("truetype");
-}
