--- src/styles/global.css
@@ -1,150 +1,0 @@
-@import "./fonts.css";
-
-* {
-  box-sizing: border-box;
-}
-
-:root {
-  --sidious-red: #ff1462;
-}
-
-html,
-body,
-#PHANTOM {
-  width: 100vw;
-  height: 100vh;
-  margin: 0;
-  padding: 0;
-  font-family: Arial;
-  font-size: 10px;
-  background-color: var(--sidious-red);
-}
-
-/* THREE CANVAS */
-canvas {
-  height: 100vh !important;
-  width: 100vw !important;
-  background: transparent;
-  position: absolute;
-  top: 0;
-}
-
-.blm-svg,
-.tara-svg,
-.aar-svg {
-  margin: 1rem;
-  box-shadow: var(--sidious-red) 5px 5px;
-}
-
-.blm-svg:hover,
-.tara-svg:hover,
-.aar-svg:hover {
-  transform: scale(1.1);
-}
-
-/* DONATION LOGO FILLS TO COUNTER PARENT'S (NAVBAR) MIX-BLEND-MODE*/
-.blm-svg rect {
-  fill: lime;
-  border-radius: 1rem;
-}
-
-.blm-svg path {
-  fill: var(--sidious-red);
-}
-
-.tara-svg rect {
-  fill: mintcream;
-  border-radius: 1rem;
-}
-
-.tara-svg path {
-  fill: var(--sidious-red);
-}
-
-.aar-svg rect {
-  fill: rgb(220, 101, 250);
-}
-
-.aar-svg path {
-  fill: oldlace;
-  stroke: mintcream;
-}
-
-.title-div {
-  font-size: 6vw;
-  padding: 3rem;
-  user-select: none;
-  width: 100%;
-  height: 100%;
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  letter-spacing: -0.3rem;
-  font-style: regular;
-  text-transform: uppercase;
-  color: var(--sidious-red);
-  mix-blend-mode: difference;
-  z-index: 99;
-}
-
-.title-h1 {
-  font-family: "Nikes";
-  cursor: pointer;
-}
-
-.born-and-raised-div {
-  background: var(--sidious-red);
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.born-and-raised-p {
-  margin-bottom: 0;
-  mix-blend-mode: overlay;
-  pointer-events: none;
-  user-select: none;
-}
-
-.desert-p {
-  margin-top: 0;
-  text-align: center;
-  mix-blend-mode: overlay;
-  font-family: Flag;
-  font-size: 14vw;
-  user-select: none;
-}
-
-.scorpion-div {
-  mix-blend-mode: multiply;
-  background-image: url("../assets/images/scorpion.png");
-  background-size: contain;
-  background-repeat: no-repeat;
-  font-family: NeueMachinaUltrabold;
-  margin: 4rem;
-  padding: 2rem;
-  font-size: 8vw;
-  position: relative;
-  width: 80%;
-  cursor: crosshair;
-}
-
-.xicuahua-img {
-  position: absolute;
-  top: 0;
-  right: 5%;
-  mix-blend-mode: difference;
-  width: 40%;
-  pointer-events: none;
-}
-
-/* MEDIA QUERIES */
-
-@media only screen and (device-width: 375px) and (device-height: 812px) {
-  /* IPHONE X */
-  .title {
-    letter-spacing: -0.1rem;
-  }
-}
