+++ webpack.config.js
@@ -3,18 +3,17 @@ const HtmlWebpackPlugin = require("html-webpack-plugin");
 const { CleanWebpackPlugin } = require("clean-webpack-plugin");
 const UglifyJsPlugin = require("uglifyjs-webpack-plugin");
 const CopyPlugin = require("copy-webpack-plugin");
-// const BundleAnalyzerPlugin = require("webpack-bundle-analyzer").BundleAnalyzerPlugin;
 
 module.exports = {
   entry: {
-    app: "./src/MOTHERBOARD.ts"
+    app: "./src/index.tsx"
   },
   resolve: {
     alias: {
       // Forward all three imports to our exports file
       three$: path.resolve("./src/three.x.ts")
     },
-    extensions: [".ts", ".js"]
+    extensions: [".tsx", ".ts", ".js"]
   },
   mode: "development",
   output: {
@@ -35,7 +34,6 @@ module.exports = {
       template: "src/index.html"
     }),
     new CleanWebpackPlugin(),
-    // new BundleAnalyzerPlugin(),
     new CopyPlugin({
       patterns: [{ from: "static", to: "static" }]
     })
@@ -64,7 +62,7 @@ module.exports = {
       },
       {
         test: /\.(svg)$/,
-        use: ["file-loader"]
+        use: ["svg-react-loader"]
       },
       {
         test: /\.(woff|woff2|eot|ttf|otf)$/,
